+++ src/main/java/in/koreatech/payment/service/TemporaryPaymentService.java
@@ -6,20 +6,15 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import in.koreatech.koin.domain.order.cart.model.Cart;
-import in.koreatech.koin.domain.order.cart.model.CartMenuItem;
-import in.koreatech.koin.domain.order.cart.model.CartMenuItemOption;
 import in.koreatech.koin.domain.order.cart.repository.CartRepository;
-import in.koreatech.koin.domain.order.shop.model.entity.menu.OrderableShopMenuOption;
-import in.koreatech.koin.domain.order.shop.model.entity.menu.OrderableShopMenuPrice;
 import in.koreatech.koin.domain.order.shop.model.entity.shop.OrderableShop;
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.payment.dto.response.TemporaryPaymentResponse;
 import in.koreatech.payment.gateway.pg.PaymentGatewayService;
+import in.koreatech.payment.mapper.TemporaryMenuItemsMapper;
 import in.koreatech.payment.model.domain.DeliveryPaymentInfo;
 import in.koreatech.payment.model.domain.TakeoutPaymentInfo;
 import in.koreatech.payment.model.domain.TemporaryMenuItems;
-import in.koreatech.payment.model.domain.TemporaryMenuOption;
-import in.koreatech.payment.model.domain.TemporaryMenuPrice;
 import in.koreatech.payment.model.redis.TemporaryPayment;
 import in.koreatech.payment.repository.redis.TemporaryPaymentRedisRepository;
 import lombok.RequiredArgsConstructor;
@@ -32,13 +27,14 @@ public class TemporaryPaymentService {
     private final CartRepository cartRepository;
     private final PaymentGatewayService paymentGatewayService;
     private final TemporaryPaymentRedisRepository temporaryPaymentRedisRepository;
+    private final TemporaryMenuItemsMapper temporaryMenuItemsMapper;
 
     @Transactional
     public TemporaryPaymentResponse createDeliveryPayment(User user, DeliveryPaymentInfo deliveryPaymentInfo) {
         Cart cart = cartRepository.getCartByUserId(user.getId());
         OrderableShop orderableShop = cart.getOrderableShop();
 
-        List<TemporaryMenuItems> temporaryMenuItems = createTemporaryMenuItemsFrom(cart);
+        List<TemporaryMenuItems> temporaryMenuItems = temporaryMenuItemsMapper.fromCart(cart);
         int totalProductPrice = cart.calculateItemsAmount();
         int deliveryFee = orderableShop.calculateDeliveryFee(totalProductPrice);
         int finalAmount = totalProductPrice + deliveryFee;
@@ -70,7 +66,7 @@ public class TemporaryPaymentService {
         Cart cart = cartRepository.getCartByUserId(user.getId());
         OrderableShop orderableShop = cart.getOrderableShop();
 
-        List<TemporaryMenuItems> temporaryMenuItems = createTemporaryMenuItemsFrom(cart);
+        List<TemporaryMenuItems> temporaryMenuItems = temporaryMenuItemsMapper.fromCart(cart);
         int totalProductPrice = cart.calculateItemsAmount();
         int finalAmount = totalProductPrice;
 
@@ -91,39 +87,5 @@ public class TemporaryPaymentService {
 
         temporaryPaymentRedisRepository.save(takeoutEntity);
         return TemporaryPaymentResponse.of(pgOrderId);
-    }
-
-    public List<TemporaryMenuItems> createTemporaryMenuItemsFrom(Cart cart) {
-        return cart.getCartMenuItems().stream()
-            .map(this::fromCartMenuItem)
-            .toList();
-    }
-
-    private TemporaryMenuItems fromCartMenuItem(CartMenuItem cartMenuItem) {
-        List<TemporaryMenuOption> options = cartMenuItem.getCartMenuItemOptions().stream()
-            .map(this::fromCartMenuItemOption)
-            .toList();
-
-        OrderableShopMenuPrice price = cartMenuItem.getOrderableShopMenuPrice();
-
-        return new TemporaryMenuItems(
-            cartMenuItem.getOrderableShopMenu().getName(),
-            cartMenuItem.getQuantity(),
-            cartMenuItem.calculateTotalAmount(),
-            new TemporaryMenuPrice(price.getName(), price.getPrice()),
-            options
-        );
-    }
-
-    private TemporaryMenuOption fromCartMenuItemOption(CartMenuItemOption option) {
-        OrderableShopMenuOption shopOption = option.getOrderableShopMenuOption();
-        String optionGroupName = shopOption.getOptionGroup().getName();
-
-        return new TemporaryMenuOption(
-            optionGroupName,
-            option.getOptionName(),
-            option.getQuantity(),
-            option.getOptionPrice()
-        );
     }
 }
